requirements.txt
watchdog
openpyxl
python-dotenv
pyinstaller